Provider better compatibility with oidc-providers:
- Remove the requirement to have the email as a claim in the JWT access
token: this is not required the specification and is not supported by a
number of OIDC providers. Normally, the Id token contains such claims.

So for now by default we add the `<username>@<username>` email when
there is no email in the JWT token. However this should be done in a
better way, possibly letting the email retrieval be done in the profile
retrieval step in this case (in `KeycloakProfileDao`).

- Explicitely specify the the `openid email profile` scope when
requesting the access token. Because OIDC providers, when answering to
the `userInfo` endpoint, are expected to return claims that corresponds
to the scopes of the access token. So if an access token has the
`openid` scope only, the `userinfo` might return no claim at all
(according to the specification).

Until now it was working since keycloak allows adding claims to the
returned tokens anyway.

- Allow supporting fixed redirect Uris: most OIDC providers support
having a list of redirect URIs to come back to after the authorization
step.  But these authorized Uris don't necessarily support wildcards or
prefix. Che doesn't support this currently, and this commit introduce 2
fixed callback HTML pages that redirect to the Dashboard / IDE URL of
the final page we want to come back to after authentication. This makes
Che compatible with more OIDC providers

Here again, some work has to be done to make the corresponding code
cleaner, more general, but this commit gives the directions to have this
working.

Signed-off-by: David Festal <dfestal@redhat.com>